feat: Регистрация и сохранение токена работают

+++ app/src/main/java/com/eltex/chat/data/models/Me.kt
@@ -11,5 +11,5 @@ data class Me(
     val status: String,
     val statusConnection: String,
     val username: String,
-    val utcOffset: Int
+    val utcOffset: Float,
 )
